+++ activiti-cloud-events-adapter-service/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  org.activiti.cloud.events.adapter.ActivitiCloudEventsAdapterAutoConfiguration
+  org.activiti.cloud.events.adapter.jms.JmsConfig
