+++ activiti-cloud-starter-events-adapter/pom.xml
@@ -10,7 +10,12 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>activiti-cloud-starter-events-adapter</artifactId>
-    <name>Activiti Cloud Events Adapter :: Starter :: Events Adapter</name>
+    <name>Activiti Cloud Events Adapter :: Starter</name>
+
+    <properties>
+        <activiti-cloud-events-adapter-service.version>0.0.1-SNAPSHOT</activiti-cloud-events-adapter-service.version>
+        <dependency.activemq-junit.version>5.15.5</dependency.activemq-junit.version>
+    </properties>
 
     <dependencies>
         <dependency>
@@ -20,7 +25,28 @@
         <dependency>
             <groupId>org.activiti.cloud</groupId>
             <artifactId>activiti-cloud-events-adapter-service</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
+            <version>${activiti-cloud-events-adapter-service.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-stream-test-support</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.activemq.tooling</groupId>
+            <artifactId>activemq-junit</artifactId>
+            <version>${dependency.activemq-junit.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.awaitility</groupId>
+            <artifactId>awaitility</artifactId>
+            <scope>test</scope>
         </dependency>
     </dependencies>
