+++ activiti-cloud-starter-events-adapter/src/test/java/org/activiti/cloud/events/adapter/ActivitiCloudEventsAdapterTest.java
@@ -15,11 +15,6 @@
  */
 package org.activiti.cloud.events.adapter;
 
-import static org.junit.Assert.assertEquals;
-
-import java.util.List;
-import java.util.concurrent.TimeUnit;
-
 import org.activiti.cloud.events.adapter.streams.EventsAdapterConsumerChannel;
 import org.activiti.cloud.events.adapter.util.EventReceiverUtil;
 import org.activiti.cloud.events.adapter.util.TestUtil;
@@ -37,6 +32,11 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.integration.channel.AbstractMessageChannel;
 import org.springframework.messaging.support.GenericMessage;
 import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+
+import static org.awaitility.Awaitility.await;
+import static org.junit.Assert.assertEquals;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
@@ -76,21 +76,22 @@ public class ActivitiCloudEventsAdapterTest {
         final String payload = getEventResourceAsString("aggregated-processStarted.json");
         adapterConsumer.send(new GenericMessage<>(payload));
         // wait for a few seconds to get all the messages
-        TimeUnit.SECONDS.sleep(5);
-        List<String> result = receiverUtil.getEvents();
-        assertEquals("11 events must have been sent.", 11, result.size());
-        //Checked the received events
-        checkExpectedJson(getEventResourceAsString("public-processCreated-01.json"), result.get(0));
-        checkExpectedJson(getEventResourceAsString("public-variableCreated-02.json"), result.get(1));
-        checkExpectedJson(getEventResourceAsString("public-variableCreated-03.json"), result.get(2));
-        checkExpectedJson(getEventResourceAsString("public-variableCreated-04.json"), result.get(3));
-        checkExpectedJson(getEventResourceAsString("public-processStarted-05.json"), result.get(4));
-        checkExpectedJson(getEventResourceAsString("public-activityStarted-06.json"), result.get(5));
-        checkExpectedJson(getEventResourceAsString("public-activityCompleted-07.json"), result.get(6));
-        checkExpectedJson(getEventResourceAsString("public-sequenceFlowTaken-08.json"), result.get(7));
-        checkExpectedJson(getEventResourceAsString("public-activityStarted-09.json"), result.get(8));
-        checkExpectedJson(getEventResourceAsString("public-taskCandidateGroupAdded-10.json"), result.get(9));
-        checkExpectedJson(getEventResourceAsString("public-taskCreated-11.json"), result.get(10));
+        await().untilAsserted(() -> {
+            List<String> result = receiverUtil.getEvents();
+            assertEquals("11 events must have been sent.", 11, result.size());
+            //Checked the received events
+            checkExpectedJson(getEventResourceAsString("public-processCreated-01.json"), result.get(0));
+            checkExpectedJson(getEventResourceAsString("public-variableCreated-02.json"), result.get(1));
+            checkExpectedJson(getEventResourceAsString("public-variableCreated-03.json"), result.get(2));
+            checkExpectedJson(getEventResourceAsString("public-variableCreated-04.json"), result.get(3));
+            checkExpectedJson(getEventResourceAsString("public-processStarted-05.json"), result.get(4));
+            checkExpectedJson(getEventResourceAsString("public-activityStarted-06.json"), result.get(5));
+            checkExpectedJson(getEventResourceAsString("public-activityCompleted-07.json"), result.get(6));
+            checkExpectedJson(getEventResourceAsString("public-sequenceFlowTaken-08.json"), result.get(7));
+            checkExpectedJson(getEventResourceAsString("public-activityStarted-09.json"), result.get(8));
+            checkExpectedJson(getEventResourceAsString("public-taskCandidateGroupAdded-10.json"), result.get(9));
+            checkExpectedJson(getEventResourceAsString("public-taskCreated-11.json"), result.get(10));
+        });
     }
 
     @Test
@@ -100,11 +101,12 @@ public class ActivitiCloudEventsAdapterTest {
         final String payload = getEventResourceAsString("aggregated-taskAssigned.json");
         adapterConsumer.send(new GenericMessage<>(payload));
         // wait for a few seconds to get all the messages
-        TimeUnit.SECONDS.sleep(3);
-        List<String> result = receiverUtil.getEvents();
-        assertEquals("1 event must have been sent.", 1, result.size());
-        //Checked the received events
-        checkExpectedJson(getEventResourceAsString("public-taskAssigned.json"), result.get(0));
+        await().untilAsserted(()->{
+            List<String> result = receiverUtil.getEvents();
+            assertEquals("1 event must have been sent.", 1, result.size());
+            //Checked the received events
+            checkExpectedJson(getEventResourceAsString("public-taskAssigned.json"), result.get(0));
+        });
     }
 
     private void checkExpectedJson(String expectedJsonBody, String actualJsonBody) throws Exception
