+++ activiti-cloud-starter-events-adapter/src/test/java/org/activiti/cloud/events/adapter/util/EventReceiverUtilConfig.java
