+++ activiti-cloud-starter-events-adapter/src/test/resources/application.properties
