+++ activiti-cloud-starter-events-adapter/src/test/resources/events/aggregated-processStarted.json
