+++ activiti-cloud-starter-events-adapter/src/test/resources/events/aggregated-taskAssigned.json
