+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-activityCompleted-07.json
