+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-activityStarted-06.json
