+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-activityStarted-09.json
