+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-processCreated-01.json
