+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-processStarted-05.json
