+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-sequenceFlowTaken-08.json
