+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-taskAssigned.json
