+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-taskCandidateGroupAdded-10.json
