+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-taskCreated-11.json
