+++ activiti-cloud-starter-events-adapter/src/test/resources/events/public-variableCreated-02.json
